Refactor: Renamed main file and updated GeoBasis_Loader import in __init__.py
