+++ __init__.py
@@ -1,3 +1,3 @@
-from .GeoBasis_Loader_main import GeoBasis_Loader
+from .main import GeoBasis_Loader
 def classFactory(iface):
     return GeoBasis_Loader(iface)
